Act093_Main.kt Frg_Serial_Edit.java    - correção de formatação de valor de medição e ultimo ciclo. @#$ Act093State.kt InfoSerialModel.kt     - Correção no tipo de valores. @#$ StringHelper.kt     - Tratamento de null para string que são exibidas para usuário. @#$ ToolBox_Inf.java     - Duplicação de metodo para double. @#$

+++ app/src/main/java/com/namoadigital/prj001/ui/act093/util/Act093State.kt
@@ -15,9 +15,9 @@ data class Act093State(
         val trackings: String? = null,
         val infoAdd: String? = null,
         val value_suffix: String? = null,
-        val last_measure_value: Int? = null,
+        val last_measure_value: Double? = null,
         val last_measure_date: String? = null,
-        val last_cycle_value: Int? = null,
+        val last_cycle_value: Float? = null,
         val last_cycle_date: String? = null,
         val lastUpdateSerial: String? = null,
     )
